requirements.txt
# Core Requirements
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# CLI Enhancement (optional but recommended)
rich>=13.5.0
colorama>=0.4.6
tqdm>=4.66.0

# Configuration
PyYAML>=6.0

# For Excel export support (optional)
openpyxl>=3.1.2
pandas>=2.0.0

# Development tools (optional)
pytest>=7.4.0
black>=23.7.0
pylint>=2.17.0

# GUI requirements (tkinter comes with Python)
# tkinter is included with most Python installations
